code_ref/alphagenome-main/pyproject.toml
[build-system]
build-backend = 'hatchling.build'
requires = ['hatchling', 'grpcio-tools<=1.67.1', 'importlib-resources']


[project]
name = 'alphagenome'
description = 'A Python SDK for interacting and visualizing genomic models.'
readme = 'README.md'
dynamic = ['version']
license = { file = 'LICENSE' }
requires-python = '>=3.10'
authors = [
    {name = 'Google LLC'},
    {email = 'alphagenome@google.com'},
]
keywords = [
    'python',
    'machine learning',
    'genomics'
]
classifiers=[
    'Development Status :: 4 - Beta',
    'Environment :: Console',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
]
dependencies=[
    # keep-sorted start
    'absl-py',
    'anndata',
    'grpcio>=1.67.1',
    'immutabledict',
    'intervaltree',
    'jaxtyping',
    'matplotlib',
    'ml_dtypes',
    'numpy',
    'pandas',
    'protobuf>=5.28.3',
    'pyarrow',
    'scipy',
    'seaborn',
    'tqdm',
    'typeguard',
    'typing_extensions',
    'zstandard',
    # keep-sorted end
]

[project.urls]
Repository = 'https://github.com/google-deepmind/alphagenome'
Documentation = 'https://www.alphagenomedocs.com/'

[project.optional-dependencies]
dev = [
  'hatch',
]
docs = [
    'ipykernel',
    'ipython',
    'myst-nb',
    'sphinx>=5.0',
    'sphinx-autodoc-typehints',
    'sphinx-book-theme',
    'sphinx-copybutton',
    'sphinx-remove-toctrees',
    'sphinx-design',
    'sphinxcontrib-bibtex>=1.0.0',
]
scripts = [
    'absl-py',
    'alphagenome',
    'pandas',
    'pyarrow',
    'pyranges',
]

# Calls hatch_build.py to generate Python bindings for protos.
[tool.hatch.build.hooks.custom]

[tool.hatch.version]
path = 'src/alphagenome/__init__.py'

[tool.hatch.envs.default]
installer = 'uv'

[tool.setuptools.packages.find]
include = ['README.md', 'LICENSE']
exclude = ['*_test.py', 'examples']

[tool.hatch.envs.hatch-test]
default-args = []
extra-dependencies=['google-benchmark', 'pyranges', 'typeguard==2.13.3']
parallel = true


[tool.hatch.envs.hatch-test.env-vars]
MPLBACKEND = 'agg'

[[tool.hatch.envs.hatch-test.matrix]]
# Use hatch test --all to run tests on all supported Python versions.
python = ['3.13', '3.12', '3.11', '3.10']

[tool.hatch.envs.check]
dependencies = [
  'nbconvert',
  'pyink>=24.3.0',
  'pylint>=2.6.0',
]
# Do not install dependencies for the check environment.
detached = true

[tool.hatch.envs.check.scripts]
format = 'pyink . --check --diff'
lint = 'pylint .'
notebook_lint = 'jupyter nbconvert colabs/*.ipynb --to html'
all = [
  'format',
  'lint',
  'notebook_lint',
]

[tool.pyink]
# Formatting configuration to follow Google style-guide
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true
exclude = 'src/alphagenome/protos'
